requirements.txt
attrs==22.1.0
click==8.1.3
coverage==6.5.0
flask[async]==2.2.2
Flask-SQLAlchemy==3.0.2
flask-swagger==0.2.14
greenlet==2.0.1
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
PyMySQL==1.0.2
PyMySQL[rsa]==1.0.2
pluggy==1.0.0
py==1.11.0
PyDispatcher==2.0.6
pyparsing==3.0.9
pytest==7.1.3
PyYAML==6.0
SQLAlchemy==1.4.44
tomli==2.0.1
Werkzeug==2.2.2
aiopulsar-py==0.1.0a0
pulsar-client==2.10.1
pulsar-client[avro]==2.10.1
fastapi==0.89.1
uvicorn[standard]==0.20.0
requests==2.28.2